refactor: rename files and improve code organization
- Rename emoji_flags.go to countries.go
- Rename test files to countries_test.go and countries_bench_test.go
- Create utils.go with utility functions
- Extract helper functions for better readability:
  - codeToFlag/flagToCode for flag conversion
  - isFlagEmoji for input validation
  - lookupAlpha2By* for code lookups
  - fuzzyMatchCode/fuzzyMatchName for fuzzy matching
  - buildCountryInfo for result construction
- Reduce complexity in GetCountryInfo by using extracted helpers
